% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_conditions.R
\name{extract_conditions}
\alias{extract_conditions}
\title{Extract Conditions}
\usage{
extract_conditions(trial_path)
}
\arguments{
\item{trial_path}{An absolute or relative path to the XML file}
}
\value{
Returns a tidy dataframe with the basic trial metadata (a trial's NCT
  number, trial status, phase and title) with the MESH terms associated to
  the trial.
}
\description{
Extract the MESH condtions assigned to the trials. CAUTION:  The MeSH terms
are assigned with an imperfect algorithm.
}
