% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_trial_duration.R
\name{extract_date}
\alias{extract_date}
\title{Extract Date}
\usage{
extract_date(trial_path)
}
\arguments{
\item{trial_path}{An absolute or relative path to the XML file}
}
\value{
Returns a tidy dataframe with the basic trial metadata (a trial's NCT number,
trial status, phase and title) with the start and end date of the trial.
}
\description{
Function to extract the start and end date of the trial.
}
