% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_eligibility.R
\name{extract_eligibility}
\alias{extract_eligibility}
\title{Extract Eligibility}
\usage{
extract_eligibility(trial_path)
}
\arguments{
\item{trial_path}{An absolute or relative path to the XML file}
}
\value{
Returns a tidy dataframe with the basic trial metadata (a trial's NCT
  number, trial status, phase and title) with the inclusion and exclusion
  criteria.
}
\description{
Function extraccts the eligibility criteria from the trials.
}
