% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_locations.R
\name{extract_locations}
\alias{extract_locations}
\title{Extract locations}
\usage{
extract_locations(trial_path)
}
\arguments{
\item{trial_path}{An absolute or relative path to the XML file}
}
\value{

}
\description{
Extract the locations of the trials
}
