% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_outcomes.R
\name{extract_outcome}
\alias{extract_outcome}
\title{Extract Outcomes}
\usage{
extract_outcome(trial_path, primary_outcome = TRUE, secondary_outcome = TRUE)
}
\arguments{
\item{trial_path}{An absolute or relative path to the XML file}

\item{primary_outcome}{Logical. If TRUE the primary outcome is extracted.}

\item{secondary_outcome}{Logical. If TRUE the secondary outcome is extracted.}
}
\value{
Returns a tidy dataframe with the basic trial metadata (a trial's NCT number,
trial status, phase and title) with the primary and/or secondary outcome.
}
\description{
Function to extract the primary and secondary outcomes.
}
\examples{

ct_files <- list.files(path = "Benchmark_data/", pattern = "\\\\.xml", recursive = TRUE, full.names = TRUE)
k <- lapply(X = ct_files, FUN = extract_outcome)
k <- dplyr::bind_rows(k)
View(k)

}
