% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_termination_reasons.R
\name{extract_termination_reason}
\alias{extract_termination_reason}
\title{Extract Termination Reason}
\usage{
extract_termination_reason(trial_path)
}
\arguments{
\item{trial_path}{An absolute or relative path to the XML file}
}
\value{
Returns a tidy dataframe with the basic trial metadata (a trial's NCT number,
trial status, phase and title) with the reasons why the trial stopped.
}
\description{
Extract the resons why a trial was terminated, if provided.
}
\examples{

ct_files <- list.files(path = "Benchmark_data/", pattern = "\\\\.xml", recursive = TRUE, full.names = TRUE)
k <- lapply(X = ct_files, FUN = extract_termination_reason)
k <- dplyr::bind_rows(k)
View(k)

}
