% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_locations.R
\name{extract_locations}
\alias{extract_locations}
\title{Extract locations}
\usage{
extract_locations(trial_path)
}
\arguments{
\item{trial_path}{}
}
\value{

}
\description{
Extract the locations of the trials
}
\examples{

ct_files <- list.files(path = "Benchmark_data/", pattern = "\\\\.xml", recursive = TRUE, full.names = TRUE)
k <- lapply(X = ct_files, FUN = extract_locations)
k <- dplyr::bind_rows(k)
View(k)

}
