% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_outcomes.R
\name{extract_outcome}
\alias{extract_outcome}
\title{Extract Outcomes}
\usage{
extract_outcome(trial_path, outcome = c("primary", "secondary"))
}
\arguments{
\item{trial_path}{An absolute or relative path to the XML file}

\item{outcome}{A character string with the outcome type (primary or secondary) that should be extracted. Default extracts both.}
}
\value{
Returns a tidy dataframe with the basic trial metadata (a trial's NCT number,
trial status, phase and title) with the primary and/or secondary outcome.
}
\description{
Function to extract the primary and secondary outcomes.
}
