% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Internal_housekeeping.R
\name{Housekeeping}
\alias{Housekeeping}
\title{Housekeeping}
\usage{
Housekeeping(trial_path)
}
\arguments{
\item{trial_path}{An absolute or relative path to the XML file}
}
\value{
Returns a dataframe with a trial's NCT number, trial status, phase
  and title. 1 row per trial
}
\description{
Collect basic metadata from a trial such as
\itemize{
  \item NCT number
  \item Trial Status
  \item Phase
  \item Brief Title
}
}
\examples{
ct_files <- example_file <- system.file("extdat", "NCT00160147.xml", package = "TidyTrials")
Housekeeping(ct_files)
}
